refactor(services): Move start_with_new_p_c
Move the function and move the printing of names to the caller so that
the same functionality can be used by `start` and `restart`.
